tools/testing/connection/python-redis/redis-py/requirements.txt
redis>=4.0.0
python-dotenv
